Library Management System
Library management is an application where all the information related
to books, branches, checked in books, checked out books, etc. is stored
in normalized form. User friendly GUI is provided to interact with the
application. Validations are done on all the fields. Wherever possible
tabular format is used to display the data and user is allowed to click
on the respective rows to make further manipulations to the data- this
is done to keep the interaction with the application user friendly.